Test_LLM/agent_backend/src/requirements.txt
pip
autogen==0.7.3
azureml-inference-server-http
azure-keyvault-secrets==4.8.0         # 4.9.0 may conflict with azure-cli
brotli==1.0.9
cloudpickle==2.2.1                     # Compatible with AzureML (under 3.0.0)
colorama==0.4.6
cython==3.0.10
dataclasses==0.6
h2==4.1.0
httpx>=0.27.2
langchain==0.2.17
langchain-openai==0.1.25
mlflow==2.20.2
numpy==1.23.5
openai>=1.12.0                         # Keep compatible with langchain-openai
pandas==1.3.5                          # AzureML-compatible
pandasql==0.7.3
psutil==5.9.3
python-dotenv
rich==13.8.1
sqlalchemy==1.4.46                     # v2 may break legacy pandasql/langchain
tornado==6.4.1
uuid==1.30
